upd: creando una mejor estructura para poder tener una mejor organización así mismo creando nuevos archivos
