+++ index.html
@@ -11,7 +11,7 @@
 
 
     <div id="root"></div>
-    <script type="module" src="/src/main.jsx"></script>
+    <script type="module" src="/src/index.jsx"></script>
 
   </body>
 </html>
