--- src/Nav/Main.html
@@ -1,59 +1,0 @@
-<!DOCTYPE html>
-<html lang="es">
-  <head>
-    <title>Inicio</title>
-    <style>
-      .tarjeta {
-        width: 30%; 
-        margin: 1%;
-        float: left; 
-        text-align: center; 
-        box-shadow: 0 4px 8px 0 rgba(0,0,0,0.2); 
-        transition: 0.3s; 
-      }
-      .tarjeta:hover {
-        box-shadow: 0 8px 16px 0 rgba(0,0,0,0.2); 
-      }
-      .container {
-        padding: 2px 16px; 
-      }
-    </style>
-  </head>
-  <body>
-    <nav>
-      <ul>
-        <li><a href="/">Inicio</a></li>
-        <li><a href="/">Control de inventario</a></li>
-        <li><a href="/">Usuarios</a></li>
-        <li><a href="/">Proveedores</a></li>
-        <li><a href="/">Clientes</a></li>
-        <li><a href="/">Materiales</a></li>
-      </ul>
-    </nav>
-    <h1>Bienvenido</h1>
-    <h4>Sistema de control de inventario de radio y television hidalgo</h4>
-    <!-- Tarjetas -->
-    <div class="tarjeta" id="tarjeta1" style="cursor: help;">
-      <div class="container">
-        <h4><b>Tarjeta 1</b></h4> 
-        <p>Descripción de la tarjeta 1</p> 
-      </div>
-    </div>
-    <div class="tarjeta" id="tarjeta2">
-      <div class="container">
-        <h4><b>Tarjeta 2</b></h4> 
-        <p>Descripción de la tarjeta 2</p> 
-      </div>
-    </div>
-    <div class="tarjeta" id="tarjeta3">
-      <!-- Envolver el contenido de la tarjeta en un enlace -->
-      <a href="/public/nav/Prueba.html" style="text-decoration: none; color: black;">
-        <div class="container">
-          <h4><b>Tarjeta 3</b></h4>
-          <p>Descripción de la tarjeta 3</p>
-        </div>
-      </a>
-    </div>
-    </div>
-  </body>
-</html>
