+++ src/views/home.jsx
@@ -1,0 +1,13 @@
+import React from 'react';
+
+function Home() {
+  return (
+    <div>
+      <h1>Bienvenido al Inicio</h1>
+      <p>Esta es la página de inicio de la aplicación.</p>
+      {/* Aquí puedes agregar más contenido o funcionalidad según lo necesites */}
+    </div>
+  );
+}
+
+export default Home;
